+++ MajstorHUB-Back/MajstorHUB/Services/FirmaService/IFirmaService.cs
@@ -1,4 +1,4 @@
-﻿namespace MajstorHUB.Services;
+﻿namespace MajstorHUB.Services.FirmaService;
 
 public interface IFirmaService
 {
