+++ MajstorHUB-Back/MajstorHUB/Services/MajstorService/IMajstorService.cs
@@ -1,4 +1,4 @@
-﻿namespace MajstorHUB.Services;
+﻿namespace MajstorHUB.Services.MajstorService;
 
 public interface IMajstorService
 {
