+++ MajstorHUB-Back/MajstorHUB/Services/MajstorService/MajstorService.cs
@@ -1,4 +1,4 @@
-﻿namespace MajstorHUB.Services;
+﻿namespace MajstorHUB.Services.MajstorService;
 
 public class MajstorService : IMajstorService
 {
