+++ MajstorHUB-Back/MajstorHUB/Services/OglasService/IOglasService.cs
@@ -1,0 +1,11 @@
+﻿namespace MajstorHUB.Services.OglasService;
+
+public interface IOglasService
+{
+    Task<List<Oglas>> GetAll();
+    Task<Oglas> GetById(string id);
+    Task<List<Oglas>> GetByKorisnik(string korisnikId);
+    Task Create(Oglas oglas);
+    Task Update(string id, Oglas oglas);
+    Task Delete(string id);
+}
